The Final Interpreter Version
I did not submit this as I did not want to hinder the grade we would get for the original submission. Instead, I took this as a personal challenge to make the SCL interpreter accept both if else statements and getline statements to read from the console. You can try it by using the java compiler in the command line or you can recompile the source files in an IDE if you want.